Reorganize player module
